+++ sentry_sdk/integrations/ariadne.py
