+++ tests/integrations/ariadne/test_ariadne.py
@@ -1,7 +1,7 @@
 import pytest
 
 import sentry_sdk
-from sentry_sdk.integrations.graphql.ariadne import AriadneIntegration
+from sentry_sdk.integrations.ariadne import AriadneIntegration
 
 
 def test_something(sentry_init, capture_events):
