+++ tests/integrations/graphene/test_graphene.py
@@ -1,7 +1,7 @@
 import pytest
 
 import sentry_sdk
-from sentry_sdk.integrations.graphql.graphene import StrawberryIntegration
+from sentry_sdk.integrations.graphene import GrapheneIntegration
 
 
 def test_something(sentry_init, capture_events):
