+++ tests/integrations/strawberry/test_strawberry.py
@@ -1,7 +1,7 @@
 import pytest
 
 import sentry_sdk
-from sentry_sdk.integrations.graphql.graphene import GrapheneIntegration
+from sentry_sdk.integrations.strawberry import StrawberryIntegration
 
 
 def test_something(sentry_init, capture_events):
