+++ tox.ini
@@ -22,6 +22,9 @@ envlist =
     {py3.7}-aiohttp-v{3.5}
     {py3.7,py3.8,py3.9,py3.10,py3.11}-aiohttp-v{3.6}
 
+    # Ariadne
+    {py3.8,py3.9,py3.10,py3.11}-ariadne
+
     # Arq
     {py3.7,py3.8,py3.9,py3.10,py3.11}-arq
 
@@ -93,6 +96,9 @@ envlist =
     # GCP
     {py3.7}-gcp
 
+    # Graphene
+    {py3.7,py3.8,py3.9,py3.10,py3.11}-graphene
+
     # Grpc
     {py3.7,py3.8,py3.9,py3.10}-grpc-v{1.40,1.44,1.48}
     {py3.7,py3.8,py3.9,py3.10,py3.11}-grpc-v{1.54,1.56,1.58}
@@ -157,6 +163,9 @@ envlist =
     {py2.7,py3.7,py3.8,py3.9,py3.10,py3.11}-sqlalchemy-v{1.2,1.3,1.4}
     {py3.7,py3.8,py3.9,py3.10,py3.11}-sqlalchemy-v{2.0}
 
+    # Strawberry
+    {py3.8,py3.9,py3.10,py3.11}-strawberry
+
     # Tornado
     {py3.7,py3.8,py3.9}-tornado-v{5}
     {py3.7,py3.8,py3.9,py3.10,py3.11}-tornado-v{6}
@@ -184,6 +193,9 @@ deps =
     aiohttp-v3.4: aiohttp>=3.4.0,<3.5.0
     aiohttp-v3.5: aiohttp>=3.5.0,<3.6.0
     aiohttp: pytest-aiohttp
+
+    # Ariadne
+    ariadne: ariadne>=0.20
 
     # Arq
     arq: arq>=0.23.0
@@ -316,6 +328,9 @@ deps =
     # for justification why greenlet is pinned here
     py3.5-gevent: greenlet==0.4.17
     {py2.7,py3.6,py3.7,py3.8,py3.9,py3.10,py3.11}-gevent: gevent>=22.10.0, <22.11.0
+
+    # Graphene
+    graphene: graphene>=3.3
 
     # Grpc
     grpc-v1.40: grpcio-tools>=1.40.0,<1.41.0
@@ -457,6 +472,9 @@ deps =
     sqlalchemy-v1.4: sqlalchemy>=1.4,<2.0
     sqlalchemy-v2.0: sqlalchemy>=2.0,<2.1
 
+    # Strawberry
+    strawberry-graphql: strawberry-graphql>=0.208
+
     # Tornado
     tornado-v5: tornado>=5,<6
     tornado-v6: tornado>=6.0a1
@@ -473,6 +491,7 @@ setenv =
     PYTHONDONTWRITEBYTECODE=1
     common: TESTPATH=tests
     aiohttp: TESTPATH=tests/integrations/aiohttp
+    ariadne: TESTPATH=tests/integrations/ariadne
     arq: TESTPATH=tests/integrations/arq
     asgi: TESTPATH=tests/integrations/asgi
     asyncpg: TESTPATH=tests/integrations/asyncpg
@@ -491,6 +510,7 @@ setenv =
     # run all tests with gevent
     gevent: TESTPATH=tests
     gcp: TESTPATH=tests/integrations/gcp
+    graphene: TESTPATH=tests/integrations/graphene
     httpx: TESTPATH=tests/integrations/httpx
     huey: TESTPATH=tests/integrations/huey
     loguru: TESTPATH=tests/integrations/loguru
@@ -507,6 +527,7 @@ setenv =
     starlette: TESTPATH=tests/integrations/starlette
     starlite: TESTPATH=tests/integrations/starlite
     sqlalchemy: TESTPATH=tests/integrations/sqlalchemy
+    strawberry: TESTPATH=tests/integrations/strawberry
     tornado: TESTPATH=tests/integrations/tornado
     trytond: TESTPATH=tests/integrations/trytond
     socket: TESTPATH=tests/integrations/socket
